Fix wrong return type in args module.

+++ include/common/fmdt/args.h
@@ -7,7 +7,7 @@ int args_find_int_min_max(int argc, char** argv, const char* arg, int def, int m
 int args_find_int_min(int argc, char** argv, const char* arg, int def, int min);
 int args_find_int_max(int argc, char** argv, const char* arg, int def, int max);
 float args_find_float(int argc, char** argv, const char* arg, float def);
-int args_find_float_min_max(int argc, char** argv, const char* arg, float def, float min, float max);
-int args_find_float_min(int argc, char** argv, const char* arg, float def, float min);
-int args_find_float_max(int argc, char** argv, const char* arg, float def, float max);
+float args_find_float_min_max(int argc, char** argv, const char* arg, float def, float min, float max);
+float args_find_float_min(int argc, char** argv, const char* arg, float def, float min);
+float args_find_float_max(int argc, char** argv, const char* arg, float def, float max);
 char* args_find_char(int argc, char** argv, const char* arg, char* def);
